+++ bw.workflow.civilcriminalfingerprintsinquiry/src/main/java/sa/gov/nic/bio/bw/workflow/civilcriminalfingerprintsinquiry/FingerprintsInquiryWorkflow.java
@@ -7,7 +7,15 @@ import sa.gov.nic.bio.bw.core.workflow.AssociatedMenu;
 import sa.gov.nic.bio.bw.core.workflow.Signal;
 import sa.gov.nic.bio.bw.core.workflow.WithLookups;
 import sa.gov.nic.bio.bw.core.workflow.WizardWorkflowBase;
+import sa.gov.nic.bio.bw.workflow.civilcriminalfingerprintsinquiry.controllers.FingerprintsSourceFxController;
+import sa.gov.nic.bio.bw.workflow.civilcriminalfingerprintsinquiry.controllers.FingerprintsSourceFxController.Source;
+import sa.gov.nic.bio.bw.workflow.civilcriminalfingerprintsinquiry.controllers.PersonIdPaneFxController;
+import sa.gov.nic.bio.bw.workflow.civilcriminalfingerprintsinquiry.controllers.ScanFingerprintCardPaneFxController;
+import sa.gov.nic.bio.bw.workflow.civilcriminalfingerprintsinquiry.controllers.SpecifyFingerprintCoordinatesPaneFxController;
+import sa.gov.nic.bio.bw.workflow.civilcriminalfingerprintsinquiry.controllers.UploadNistFileFxController;
+import sa.gov.nic.bio.bw.workflow.civilcriminalfingerprintsinquiry.tasks.ExtractingDataFromNistFileWorkflowTask;
 import sa.gov.nic.bio.bw.workflow.commons.beans.ConvictedReport;
+import sa.gov.nic.bio.bw.workflow.commons.beans.DeporteeInfo;
 import sa.gov.nic.bio.bw.workflow.commons.beans.DisCriminalReport;
 import sa.gov.nic.bio.bw.workflow.commons.beans.PersonInfo;
 import sa.gov.nic.bio.bw.workflow.commons.controllers.InquiryByFingerprintsPaneFxController;
@@ -25,6 +33,7 @@ import sa.gov.nic.bio.bw.workflow.commons.tasks.ConvertFingerprintBase64ImagesTo
 import sa.gov.nic.bio.bw.workflow.commons.tasks.ConvertWsqFingerprintsToSegmentedFingerprintBase64ImagesWorkflowTask;
 import sa.gov.nic.bio.bw.workflow.commons.tasks.ConvictedReportInquiryFromDisWorkflowTask;
 import sa.gov.nic.bio.bw.workflow.commons.tasks.ConvictedReportToPersonInfoConverter;
+import sa.gov.nic.bio.bw.workflow.commons.tasks.DeporteeInfoToPersonInfoConverter;
 import sa.gov.nic.bio.bw.workflow.commons.tasks.DisCriminalReportToPersonInfoConverter;
 import sa.gov.nic.bio.bw.workflow.commons.tasks.FetchingFingerprintsWorkflowTask;
 import sa.gov.nic.bio.bw.workflow.commons.tasks.FetchingMissingFingerprintsWorkflowTask;
@@ -32,14 +41,8 @@ import sa.gov.nic.bio.bw.workflow.commons.tasks.FingerprintInquiryStatusCheckerW
 import sa.gov.nic.bio.bw.workflow.commons.tasks.FingerprintInquiryStatusCheckerWorkflowTask.Status;
 import sa.gov.nic.bio.bw.workflow.commons.tasks.FingerprintInquiryWorkflowTask;
 import sa.gov.nic.bio.bw.workflow.commons.tasks.FingerprintsWsqToFingerConverter;
+import sa.gov.nic.bio.bw.workflow.commons.tasks.GetDeporteeInfoByIdWorkflowTask;
 import sa.gov.nic.bio.bw.workflow.commons.tasks.GetPersonInfoByIdWorkflowTask;
-import sa.gov.nic.bio.bw.workflow.civilcriminalfingerprintsinquiry.controllers.FingerprintsSourceFxController;
-import sa.gov.nic.bio.bw.workflow.civilcriminalfingerprintsinquiry.controllers.FingerprintsSourceFxController.Source;
-import sa.gov.nic.bio.bw.workflow.civilcriminalfingerprintsinquiry.controllers.PersonIdPaneFxController;
-import sa.gov.nic.bio.bw.workflow.civilcriminalfingerprintsinquiry.controllers.ScanFingerprintCardPaneFxController;
-import sa.gov.nic.bio.bw.workflow.civilcriminalfingerprintsinquiry.controllers.SpecifyFingerprintCoordinatesPaneFxController;
-import sa.gov.nic.bio.bw.workflow.civilcriminalfingerprintsinquiry.controllers.UploadNistFileFxController;
-import sa.gov.nic.bio.bw.workflow.civilcriminalfingerprintsinquiry.tasks.ExtractingDataFromNistFileWorkflowTask;
 
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -266,12 +269,32 @@ public class FingerprintsInquiryWorkflow extends WizardWorkflowBase
 							
 							for(Long civilPersonId : civilPersonIds)
 							{
-								setData(GetPersonInfoByIdWorkflowTask.class, "personId", civilPersonId);
-								setData(GetPersonInfoByIdWorkflowTask.class,
-								        "returnNullResultInCaseNotFound", Boolean.TRUE);
-								executeWorkflowTask(GetPersonInfoByIdWorkflowTask.class);
-								civilPersonInfoMap.put(civilPersonId, getData(GetPersonInfoByIdWorkflowTask.class,
-								                                              "personInfo"));
+								if(civilPersonId == null) continue;
+								
+								PersonInfo personInfo;
+								
+								String sCivilPersonId = String.valueOf(civilPersonId);
+								if(sCivilPersonId.length() == 10 && sCivilPersonId.startsWith("9"))
+								{
+									setData(GetDeporteeInfoByIdWorkflowTask.class, "deporteeId",
+									        civilPersonId);
+									setData(GetDeporteeInfoByIdWorkflowTask.class,
+									        "returnNullResultInCaseNotFound", Boolean.TRUE);
+									executeWorkflowTask(GetDeporteeInfoByIdWorkflowTask.class);
+									DeporteeInfo deporteeInfo = getData(GetDeporteeInfoByIdWorkflowTask.class,
+									                                    "deporteeInfo");
+									personInfo = new DeporteeInfoToPersonInfoConverter().convert(deporteeInfo);
+								}
+								else
+								{
+									setData(GetPersonInfoByIdWorkflowTask.class, "personId", civilPersonId);
+									setData(GetPersonInfoByIdWorkflowTask.class,
+									        "returnNullResultInCaseNotFound", Boolean.TRUE);
+									executeWorkflowTask(GetPersonInfoByIdWorkflowTask.class);
+									personInfo = getData(GetPersonInfoByIdWorkflowTask.class, "personInfo");
+								}
+								
+								civilPersonInfoMap.put(civilPersonId, personInfo);
 							}
 							
 							setData(getClass(), FIELD_CIVIL_PERSON_INFO_MAP, civilPersonInfoMap);